requirements.txt
streamlit>=1.37
pandas
numpy
scikit-learn
opencv-python-headless
fpdf2
psutil
pillow
